refactor: :art: Fix pages structure and https methods
Decoupled logic from components and moved to the top level. Fix http get and post.
