+++ frontend/src/utils/endpoint.ts
@@ -1,0 +1,6 @@
+export const ENDPOINTS = {
+  BASEURL: `http://localhost:8080/api/`,
+  get BOOKS() {
+    return this.BASEURL + `books`;
+  },
+};
